Add a few functions
